Modified_payments

+++ src/main/java/com/example/matrimony/dto/UpdateSubscriptionPlanFullRequest.java
@@ -1,11 +1,13 @@
 package com.example.matrimony.dto;
 
+import java.math.BigDecimal;
+
 import com.example.matrimony.entity.DiscountType;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 public class UpdateSubscriptionPlanFullRequest {
 
-    private Long festivalPrice;
+    private BigDecimal festivalPrice;
     @JsonFormat(pattern = "yyyy-MM-dd")// null = disable festival pricing
    
     private String festivalStart; // yyyy-MM-dd'T'HH:mm
@@ -15,10 +17,10 @@ public class UpdateSubscriptionPlanFullRequest {
     private String planCode;
     private String planName;
     private Integer durationMonths;
-    private Long priceRupees;
+    private BigDecimal priceRupees;
     
     private DiscountType discountType;
-    private Long discountValue;
+    private BigDecimal discountValue;
     @JsonFormat(pattern = "yyyy-MM-dd")
     private String discountStart;
     @JsonFormat(pattern = "yyyy-MM-dd")
@@ -28,8 +30,8 @@ public class UpdateSubscriptionPlanFullRequest {
    
 
     
-    public Long getFestivalPrice() { return festivalPrice; }
-    public void setFestivalPrice(Long festivalPrice) {
+    public BigDecimal getFestivalPrice() { return festivalPrice; }
+    public void setFestivalPrice(BigDecimal festivalPrice) {
         this.festivalPrice = festivalPrice;
     }
 
@@ -60,10 +62,10 @@ public class UpdateSubscriptionPlanFullRequest {
 	public void setDurationMonths(Integer durationMonths) {
 		this.durationMonths = durationMonths;
 	}
-	public Long getPriceRupees() {
+	public BigDecimal getPriceRupees() {
 		return priceRupees;
 	}
-	public void setPriceRupees(Long priceRupees) {
+	public void setPriceRupees(BigDecimal priceRupees) {
 		this.priceRupees = priceRupees;
 	}
 	public DiscountType getDiscountType() {
@@ -72,10 +74,10 @@ public class UpdateSubscriptionPlanFullRequest {
 	public void setDiscountType(DiscountType discountType) {
 		this.discountType = discountType;
 	}
-	public Long getDiscountValue() {
+	public BigDecimal getDiscountValue() {
 		return discountValue;
 	}
-	public void setDiscountValue(Long discountValue) {
+	public void setDiscountValue(BigDecimal discountValue) {
 		this.discountValue = discountValue;
 	}
 	public String getDiscountStart() {
